Rename BoxForm to AddBoxForm, ADdBoxForm to AddBoxFormPage
