+++ src/components/AddBoxForm/DropZone/DropZone.css
